Change getAvailable... Helpers to Static
The getAvailable... functions do not depend on any instance-data. They could
be used to create configuration forms, preseting the returned values as 
options.
When doing this, it is unconvenient to having to create a full instance of
the Sermepa class and having to specify all of its attributes (they may even be
unknown at the moment).
Therefore I suggest to make this method static helpers, so they can be called
both from an instance and from the class directly.